+++ feature/settings/src/main/java/com/paradox543/malankaraorthodoxliturgica/ui/screens/AboutScreen.kt
@@ -25,9 +25,9 @@ import androidx.compose.ui.res.painterResource
 import androidx.compose.ui.text.style.TextAlign
 import androidx.compose.ui.unit.dp
 import androidx.core.net.toUri
-import com.paradox543.malankaraorthodoxliturgica.BuildConfig
-import com.paradox543.malankaraorthodoxliturgica.R
 import com.paradox543.malankaraorthodoxliturgica.core.ui.ScaffoldUiState
+import com.paradox543.malankaraorthodoxliturgica.feature.settings.BuildConfig
+import com.paradox543.malankaraorthodoxliturgica.feature.settings.R
 
 @Composable
 fun AboutScreen(